pyproject.toml
[project]
name = "BiasSphere"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    {name = "Hadidi, Hana", email = "hana.hadidi.25@ucl.ac.uk"},
    {name = "Holwerda, Wietske", email = "wietske.holwerda.25@ucl.ac.uk"},
    {name = "Lin, Grace Y.", email = "grace.lin.25@ucl.ac.uk"},
    {name = "Mothner, Zach", email = "zach.mothner.25@ucl.ac.uk"},
]
requires-python = ">=3.13,<4.0"

dependencies = [
    "fastapi>=0.121.2",
    "nicegui>=3.3.0",
    "poetry>=2.2.1",
    "pre-commit>=4.4.0",
    "pydantic>=2.12.4",
    "tool>=0.8.0",
    "uvicorn>=0.38.0",
    "dynaconf (>=3.2.12,<4.0.0)",
    "opentelemetry-sdk (>=1.38.0,<2.0.0)",
    "langchain-valyu (==0.3.4)",
    "langgraph (==1.0.1)",
    "langchain-openai (==0.3.35)",
]

[dependency-groups]
dev = [
    "detect-secrets>=1.5.0",
    "ipykernel>=7.1.0",
    "ipython>=9.7.0",
    "langchain-openai>=0.3.35",
    "langchain-valyu>=0.3.4",
    "langgraph>=1.0.1",
    "pre-commit>=4.4.0",
    "python-dotenv>=1.2.1",
    "ruff (>=0.14.5,<0.15.0)",
]

[tool.poetry]
version = "2.1.1"
packages = [{include = "src/"}]

[tool.poetry.scripts]
"app:dev" = "src.ui.main:run_app"
"api:dev" = "src.api.server:run"

[tool.poetry.group.dev.dependencies]
detect-secrets = "^1.5.0"

[tool.ruff]
line-length = 120
exclude = ["models/**", ".venv"]
